Created Functions and Support for Word Hueristics
Firstly, I moved the code for updating C, G, and Y and the code for updating the word list to their respective functions. Second, I made it that for the remaining words, it calculates the average number of remaining words given all possible answer choices; basically this should choose the word that excludes the most other words on average, but without knowledge of the actual answer.